% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ala.R
\docType{data}
\name{ala_10k}
\alias{ala_10k}
\title{ALA data| Heterogeneous | 10,000 record}
\format{
A data frame with 10000 rows and 78 variables:
\describe{
  \item{\code{uuid}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_catalogNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonConceptID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventDate}}{double COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{vernacularName}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonRank}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonRankID}}{integer COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{classs}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{genusGuid}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{character COLUMN_DESCRIPTION}
  \item{\code{speciesGuid}}{character COLUMN_DESCRIPTION}
  \item{\code{subspecies}}{character COLUMN_DESCRIPTION}
  \item{\code{subspeciesGuid}}{character COLUMN_DESCRIPTION}
  \item{\code{stateProvince}}{character COLUMN_DESCRIPTION}
  \item{\code{latitude}}{double COLUMN_DESCRIPTION}
  \item{\code{longitude}}{double COLUMN_DESCRIPTION}
  \item{\code{coordinateUncertaintyInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{year}}{integer COLUMN_DESCRIPTION}
  \item{\code{month}}{character COLUMN_DESCRIPTION}
  \item{\code{basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_locationRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{left}}{integer COLUMN_DESCRIPTION}
  \item{\code{right}}{integer COLUMN_DESCRIPTION}
  \item{\code{dataProviderUid}}{character COLUMN_DESCRIPTION}
  \item{\code{dataProviderName}}{character COLUMN_DESCRIPTION}
  \item{\code{dataResourceUid}}{character COLUMN_DESCRIPTION}
  \item{\code{dataResourceName}}{character COLUMN_DESCRIPTION}
  \item{\code{assertions}}{list COLUMN_DESCRIPTION}
  \item{\code{speciesGroups}}{list COLUMN_DESCRIPTION}
  \item{\code{geospatialKosher}}{character COLUMN_DESCRIPTION}
  \item{\code{collector}}{list COLUMN_DESCRIPTION}
  \item{\code{collectors}}{list COLUMN_DESCRIPTION}
  \item{\code{raw_scientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_vernacularName}}{character COLUMN_DESCRIPTION}
  \item{\code{latLong}}{character COLUMN_DESCRIPTION}
  \item{\code{point1}}{character COLUMN_DESCRIPTION}
  \item{\code{point01}}{character COLUMN_DESCRIPTION}
  \item{\code{point001}}{character COLUMN_DESCRIPTION}
  \item{\code{point0001}}{character COLUMN_DESCRIPTION}
  \item{\code{point00001}}{character COLUMN_DESCRIPTION}
  \item{\code{namesLsid}}{character COLUMN_DESCRIPTION}
  \item{\code{license}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_occurrenceRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_institutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_collectionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationVerificationStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceDetails}}{character COLUMN_DESCRIPTION}
  \item{\code{rights}}{character COLUMN_DESCRIPTION}
  \item{\code{eventDateEnd}}{double COLUMN_DESCRIPTION}
  \item{\code{image}}{character COLUMN_DESCRIPTION}
  \item{\code{images}}{list COLUMN_DESCRIPTION}
  \item{\code{multimedia}}{list COLUMN_DESCRIPTION}
  \item{\code{recordNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{imageUrl}}{character COLUMN_DESCRIPTION}
  \item{\code{largeImageUrl}}{character COLUMN_DESCRIPTION}
  \item{\code{smallImageUrl}}{character COLUMN_DESCRIPTION}
  \item{\code{thumbnailUrl}}{character COLUMN_DESCRIPTION}
  \item{\code{imageUrls}}{list COLUMN_DESCRIPTION}
  \item{\code{stateConservation}}{character COLUMN_DESCRIPTION}
  \item{\code{dataHubUid}}{list COLUMN_DESCRIPTION}
  \item{\code{institutionUid}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionName}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionUid}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionName}}{character COLUMN_DESCRIPTION}
  \item{\code{austConservation}}{character COLUMN_DESCRIPTION}
  \item{\code{raw_typeStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{typeStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{prov}}{character COLUMN_DESCRIPTION}
}
}
\usage{
ala_10k
}
\description{
A medium size dataset for testing
}
\details{
DETAILS
}
\keyword{datasets}
