% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_data.R
\docType{data}
\name{gbif_10k_latest}
\alias{gbif_10k_latest}
\title{gbif data with newly parsed 10,000 records (Timestamp: Tue Dec 17 18:55:27 2019)}
\format{A data frame with 10,000 rows and 113 variables:
\describe{
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{longitude}}{double COLUMN_DESCRIPTION}
  \item{\code{latitude}}{double COLUMN_DESCRIPTION}
  \item{\code{issues}}{character COLUMN_DESCRIPTION}
  \item{\code{prov}}{character COLUMN_DESCRIPTION}
  \item{\code{key}}{character COLUMN_DESCRIPTION}
  \item{\code{scientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetKey}}{character COLUMN_DESCRIPTION}
  \item{\code{publishingOrgKey}}{character COLUMN_DESCRIPTION}
  \item{\code{installationKey}}{character COLUMN_DESCRIPTION}
  \item{\code{publishingCountry}}{character COLUMN_DESCRIPTION}
  \item{\code{protocol}}{character COLUMN_DESCRIPTION}
  \item{\code{lastCrawled}}{character COLUMN_DESCRIPTION}
  \item{\code{lastParsed}}{character COLUMN_DESCRIPTION}
  \item{\code{crawlId}}{integer COLUMN_DESCRIPTION}
  \item{\code{basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{kingdomKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{phylumKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{classKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{orderKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{familyKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{genusKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{speciesKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{acceptedTaxonKey}}{integer COLUMN_DESCRIPTION}
  \item{\code{acceptedScientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{character COLUMN_DESCRIPTION}
  \item{\code{genericName}}{character COLUMN_DESCRIPTION}
  \item{\code{specificEpithet}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonRank}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonomicStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateUncertaintyInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{stateProvince}}{character COLUMN_DESCRIPTION}
  \item{\code{year}}{integer COLUMN_DESCRIPTION}
  \item{\code{month}}{integer COLUMN_DESCRIPTION}
  \item{\code{day}}{integer COLUMN_DESCRIPTION}
  \item{\code{eventDate}}{double COLUMN_DESCRIPTION}
  \item{\code{lastInterpreted}}{character COLUMN_DESCRIPTION}
  \item{\code{references}}{character COLUMN_DESCRIPTION}
  \item{\code{license}}{character COLUMN_DESCRIPTION}
  \item{\code{geodeticDatum}}{character COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{identifier}}{character COLUMN_DESCRIPTION}
  \item{\code{catalogNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{vernacularName}}{character COLUMN_DESCRIPTION}
  \item{\code{otherCatalogNumbers}}{character COLUMN_DESCRIPTION}
  \item{\code{dataGeneralizations}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonConceptID}}{character COLUMN_DESCRIPTION}
  \item{\code{locality}}{character COLUMN_DESCRIPTION}
  \item{\code{gbifID}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{character COLUMN_DESCRIPTION}
  \item{\code{individualCount}}{integer COLUMN_DESCRIPTION}
  \item{\code{eventID}}{character COLUMN_DESCRIPTION}
  \item{\code{recordedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{identifiedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{lifeStage}}{character COLUMN_DESCRIPTION}
  \item{\code{infraspecificEpithet}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionID}}{character COLUMN_DESCRIPTION}
  \item{\code{higherGeography}}{character COLUMN_DESCRIPTION}
  \item{\code{endDayOfYear}}{character COLUMN_DESCRIPTION}
  \item{\code{county}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetName}}{character COLUMN_DESCRIPTION}
  \item{\code{type}}{character COLUMN_DESCRIPTION}
  \item{\code{preparations}}{character COLUMN_DESCRIPTION}
  \item{\code{startDayOfYear}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{higherClassification}}{character COLUMN_DESCRIPTION}
  \item{\code{sex}}{character COLUMN_DESCRIPTION}
  \item{\code{elevation}}{double COLUMN_DESCRIPTION}
  \item{\code{elevationAccuracy}}{double COLUMN_DESCRIPTION}
  \item{\code{modified}}{character COLUMN_DESCRIPTION}
  \item{\code{fieldNotes}}{character COLUMN_DESCRIPTION}
  \item{\code{municipality}}{character COLUMN_DESCRIPTION}
  \item{\code{dynamicProperties}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedReferences}}{character COLUMN_DESCRIPTION}
  \item{\code{locationID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventTime}}{character COLUMN_DESCRIPTION}
  \item{\code{behavior}}{character COLUMN_DESCRIPTION}
  \item{\code{habitat}}{character COLUMN_DESCRIPTION}
  \item{\code{footprintWKT}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetID}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationVerificationStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedTaxa}}{character COLUMN_DESCRIPTION}
  \item{\code{language}}{character COLUMN_DESCRIPTION}
  \item{\code{locationAccordingTo}}{character COLUMN_DESCRIPTION}
  \item{\code{bibliographicCitation}}{character COLUMN_DESCRIPTION}
  \item{\code{organismQuantity}}{character COLUMN_DESCRIPTION}
  \item{\code{organismQuantityType}}{character COLUMN_DESCRIPTION}
  \item{\code{organismName}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimCoordinateSystem}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinatePrecision}}{double COLUMN_DESCRIPTION}
  \item{\code{dateIdentified}}{character COLUMN_DESCRIPTION}
  \item{\code{rightsHolder}}{character COLUMN_DESCRIPTION}
  \item{\code{http://unknown.org/nick}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimEventDate}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimLocality}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonID}}{character COLUMN_DESCRIPTION}
  \item{\code{http://unknown.org/occurrenceDetails}}{character COLUMN_DESCRIPTION}
  \item{\code{rights}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationID}}{character COLUMN_DESCRIPTION}
  \item{\code{informationWithheld}}{character COLUMN_DESCRIPTION}
  \item{\code{http://unknown.org/recordedByOrcid}}{character COLUMN_DESCRIPTION}
}}
\usage{
gbif_10k_latest
}
\description{
gbif data for testing data checks (via bdchecks)
}
\details{
DETAILS
}
\keyword{datasets}
