% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_branch.R
\name{get_local_branch}
\alias{get_local_branch}
\title{To get branch name}
\usage{
get_local_branch(dir_path = ".")
}
\arguments{
\item{dir_path}{a directory path of your git repository
by defult it is your current working directory

Get current branch name}
}
\value{
current chrckout branch info
}
\description{
To get branch name
}
