% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report}
\alias{get_report}
\title{To get report}
\usage{
get_report(report = TRUE)
}
\arguments{
\item{report}{pass true false}
}
\value{
prints rmarkdown report in HTML
}
\description{
To generate report as per rOpenSci
and bdverse standard
}
