% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idigbio_data.R
\docType{data}
\name{idigbio_10k_latest}
\alias{idigbio_10k_latest}
\title{idigbio data with newly parsed 10,000 records (Timestamp: Tue Dec 17 18:55:27 2019)}
\format{
A data frame with 10,000 rows and 113 variables:
\describe{
  \item{\code{associatedsequences}}{logical COLUMN_DESCRIPTION}
  \item{\code{barcodevalue}}{logical COLUMN_DESCRIPTION}
  \item{\code{basisofrecord}}{character COLUMN_DESCRIPTION}
  \item{\code{bed}}{logical COLUMN_DESCRIPTION}
  \item{\code{canonicalname}}{character COLUMN_DESCRIPTION}
  \item{\code{catalognumber}}{character COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{collectioncode}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionid}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionname}}{logical COLUMN_DESCRIPTION}
  \item{\code{collector}}{character COLUMN_DESCRIPTION}
  \item{\code{commonname}}{character COLUMN_DESCRIPTION}
  \item{\code{commonnames}}{list COLUMN_DESCRIPTION}
  \item{\code{continent}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateuncertainty}}{double COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{countrycode}}{character COLUMN_DESCRIPTION}
  \item{\code{county}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetid}}{character COLUMN_DESCRIPTION}
  \item{\code{datecollected}}{double COLUMN_DESCRIPTION}
  \item{\code{datemodified}}{character COLUMN_DESCRIPTION}
  \item{\code{dqs}}{double COLUMN_DESCRIPTION}
  \item{\code{earliestageorloweststage}}{character COLUMN_DESCRIPTION}
  \item{\code{earliesteonorlowesteonothem}}{logical COLUMN_DESCRIPTION}
  \item{\code{earliestepochorlowestseries}}{character COLUMN_DESCRIPTION}
  \item{\code{earliesteraorlowesterathem}}{character COLUMN_DESCRIPTION}
  \item{\code{earliestperiodorlowestsystem}}{character COLUMN_DESCRIPTION}
  \item{\code{etag}}{character COLUMN_DESCRIPTION}
  \item{\code{eventdate}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{fieldnumber}}{character COLUMN_DESCRIPTION}
  \item{\code{flags}}{list COLUMN_DESCRIPTION}
  \item{\code{formation}}{character COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{geologicalcontextid}}{logical COLUMN_DESCRIPTION}
  \item{\code{longitude}}{double COLUMN_DESCRIPTION}
  \item{\code{latitude}}{double COLUMN_DESCRIPTION}
  \item{\code{group}}{character COLUMN_DESCRIPTION}
  \item{\code{hasImage}}{logical COLUMN_DESCRIPTION}
  \item{\code{hasMedia}}{logical COLUMN_DESCRIPTION}
  \item{\code{highertaxon}}{character COLUMN_DESCRIPTION}
  \item{\code{highestbiostratigraphiczone}}{logical COLUMN_DESCRIPTION}
  \item{\code{individualcount}}{integer COLUMN_DESCRIPTION}
  \item{\code{infraspecificepithet}}{character COLUMN_DESCRIPTION}
  \item{\code{institutioncode}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionid}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionname}}{logical COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{latestageorhigheststage}}{logical COLUMN_DESCRIPTION}
  \item{\code{latesteonorhighesteonothem}}{logical COLUMN_DESCRIPTION}
  \item{\code{latestepochorhighestseries}}{logical COLUMN_DESCRIPTION}
  \item{\code{latesteraorhighesterathem}}{logical COLUMN_DESCRIPTION}
  \item{\code{latestperiodorhighestsystem}}{logical COLUMN_DESCRIPTION}
  \item{\code{lithostratigraphicterms}}{logical COLUMN_DESCRIPTION}
  \item{\code{locality}}{character COLUMN_DESCRIPTION}
  \item{\code{lowestbiostratigraphiczone}}{character COLUMN_DESCRIPTION}
  \item{\code{maxdepth}}{double COLUMN_DESCRIPTION}
  \item{\code{maxelevation}}{double COLUMN_DESCRIPTION}
  \item{\code{mediarecords}}{list COLUMN_DESCRIPTION}
  \item{\code{member}}{character COLUMN_DESCRIPTION}
  \item{\code{mindepth}}{double COLUMN_DESCRIPTION}
  \item{\code{minelevation}}{double COLUMN_DESCRIPTION}
  \item{\code{municipality}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceid}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{recordids}}{list COLUMN_DESCRIPTION}
  \item{\code{recordnumber}}{character COLUMN_DESCRIPTION}
  \item{\code{recordset}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{specificepithet}}{character COLUMN_DESCRIPTION}
  \item{\code{startdayofyear}}{integer COLUMN_DESCRIPTION}
  \item{\code{stateprovince}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonid}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonomicstatus}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonrank}}{character COLUMN_DESCRIPTION}
  \item{\code{typestatus}}{character COLUMN_DESCRIPTION}
  \item{\code{uuid}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimeventdate}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimlocality}}{character COLUMN_DESCRIPTION}
  \item{\code{version}}{logical COLUMN_DESCRIPTION}
  \item{\code{waterbody}}{character COLUMN_DESCRIPTION}
  \item{\code{prov}}{character COLUMN_DESCRIPTION}
}
}
\usage{
idigbio_10k_latest
}
\description{
idigbio data for testing data checks (via bdchecks)
}
\details{
DETAILS
}
\keyword{datasets}
