% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_inat.R
\docType{data}
\name{inat_10k}
\alias{inat_10k}
\title{INaturalist data| Heterogeneous | 10,000 record}
\format{
A data frame with 10000 rows and 154 variables:
\describe{
  \item{\code{quality_grade}}{character COLUMN_DESCRIPTION}
  \item{\code{time_observed_at}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon_geoprivacy}}{character COLUMN_DESCRIPTION}
  \item{\code{annotations}}{list COLUMN_DESCRIPTION}
  \item{\code{uuid}}{character COLUMN_DESCRIPTION}
  \item{\code{id}}{integer COLUMN_DESCRIPTION}
  \item{\code{cached_votes_total}}{integer COLUMN_DESCRIPTION}
  \item{\code{identifications_most_agree}}{logical COLUMN_DESCRIPTION}
  \item{\code{species_guess}}{character COLUMN_DESCRIPTION}
  \item{\code{identifications_most_disagree}}{logical COLUMN_DESCRIPTION}
  \item{\code{tags}}{character COLUMN_DESCRIPTION}
  \item{\code{positional_accuracy}}{integer COLUMN_DESCRIPTION}
  \item{\code{comments_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{site_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{created_time_zone}}{character COLUMN_DESCRIPTION}
  \item{\code{license_code}}{character COLUMN_DESCRIPTION}
  \item{\code{observed_time_zone}}{character COLUMN_DESCRIPTION}
  \item{\code{quality_metrics}}{list COLUMN_DESCRIPTION}
  \item{\code{public_positional_accuracy}}{integer COLUMN_DESCRIPTION}
  \item{\code{reviewed_by}}{list COLUMN_DESCRIPTION}
  \item{\code{oauth_application_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{flags}}{list COLUMN_DESCRIPTION}
  \item{\code{created_at}}{character COLUMN_DESCRIPTION}
  \item{\code{description}}{character COLUMN_DESCRIPTION}
  \item{\code{time_zone_offset}}{character COLUMN_DESCRIPTION}
  \item{\code{project_ids_with_curator_id}}{list COLUMN_DESCRIPTION}
  \item{\code{observed_on}}{double COLUMN_DESCRIPTION}
  \item{\code{observed_on_string}}{character COLUMN_DESCRIPTION}
  \item{\code{updated_at}}{character COLUMN_DESCRIPTION}
  \item{\code{sounds}}{list COLUMN_DESCRIPTION}
  \item{\code{place_ids}}{list COLUMN_DESCRIPTION}
  \item{\code{captive}}{logical COLUMN_DESCRIPTION}
  \item{\code{ident_taxon_ids}}{list COLUMN_DESCRIPTION}
  \item{\code{outlinks}}{list COLUMN_DESCRIPTION}
  \item{\code{faves_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{ofvs}}{list COLUMN_DESCRIPTION}
  \item{\code{num_identification_agreements}}{integer COLUMN_DESCRIPTION}
  \item{\code{comments}}{list COLUMN_DESCRIPTION}
  \item{\code{map_scale}}{integer COLUMN_DESCRIPTION}
  \item{\code{uri}}{character COLUMN_DESCRIPTION}
  \item{\code{project_ids}}{list COLUMN_DESCRIPTION}
  \item{\code{community_taxon_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{owners_identification_from_vision}}{logical COLUMN_DESCRIPTION}
  \item{\code{identifications_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{obscured}}{logical COLUMN_DESCRIPTION}
  \item{\code{num_identification_disagreements}}{integer COLUMN_DESCRIPTION}
  \item{\code{geoprivacy}}{character COLUMN_DESCRIPTION}
  \item{\code{location}}{character COLUMN_DESCRIPTION}
  \item{\code{votes}}{list COLUMN_DESCRIPTION}
  \item{\code{spam}}{logical COLUMN_DESCRIPTION}
  \item{\code{mappable}}{logical COLUMN_DESCRIPTION}
  \item{\code{identifications_some_agree}}{logical COLUMN_DESCRIPTION}
  \item{\code{project_ids_without_curator_id}}{list COLUMN_DESCRIPTION}
  \item{\code{place_guess}}{character COLUMN_DESCRIPTION}
  \item{\code{identifications}}{list COLUMN_DESCRIPTION}
  \item{\code{project_observations}}{list COLUMN_DESCRIPTION}
  \item{\code{observation_sounds}}{list COLUMN_DESCRIPTION}
  \item{\code{photos}}{list COLUMN_DESCRIPTION}
  \item{\code{observation_photos}}{list COLUMN_DESCRIPTION}
  \item{\code{faves}}{list COLUMN_DESCRIPTION}
  \item{\code{non_owner_ids}}{list COLUMN_DESCRIPTION}
  \item{\code{observed_on_details.date}}{character COLUMN_DESCRIPTION}
  \item{\code{observed_on_details.week}}{integer COLUMN_DESCRIPTION}
  \item{\code{observed_on_details.month}}{integer COLUMN_DESCRIPTION}
  \item{\code{observed_on_details.hour}}{integer COLUMN_DESCRIPTION}
  \item{\code{observed_on_details.year}}{integer COLUMN_DESCRIPTION}
  \item{\code{observed_on_details.day}}{integer COLUMN_DESCRIPTION}
  \item{\code{created_at_details.date}}{character COLUMN_DESCRIPTION}
  \item{\code{created_at_details.week}}{integer COLUMN_DESCRIPTION}
  \item{\code{created_at_details.month}}{integer COLUMN_DESCRIPTION}
  \item{\code{created_at_details.hour}}{integer COLUMN_DESCRIPTION}
  \item{\code{created_at_details.year}}{integer COLUMN_DESCRIPTION}
  \item{\code{created_at_details.day}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.is_active}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.ancestry}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.min_species_ancestry}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.endemic}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.iconic_taxon_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.min_species_taxon_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.threatened}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.rank_level}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.introduced}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.native}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.parent_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.rank}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.extinct}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.ancestor_ids}}{list COLUMN_DESCRIPTION}
  \item{\code{taxon.photos_locked}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.taxon_schemes_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.wikipedia_url}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.current_synonymous_taxon_ids}}{list COLUMN_DESCRIPTION}
  \item{\code{taxon.created_at}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.taxon_changes_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.complete_species_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.universal_search_rank}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.observations_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.atlas_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.iconic_taxon_name}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.preferred_common_name}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.complete_rank}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.flag_counts.resolved}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.flag_counts.unresolved}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.license_code}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.attribution}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.url}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.flags}}{list COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.square_url}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.medium_url}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.original_dimensions.height}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.default_photo.original_dimensions.width}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.place_id}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.source_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.user_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.authority}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.status}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.status_name}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.geoprivacy}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.iucn}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.user.id}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.user.login}}{character COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.user.icon_url}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxon.conservation_status.user.orcid}}{logical COLUMN_DESCRIPTION}
  \item{\code{preferences.prefers_community_taxon}}{logical COLUMN_DESCRIPTION}
  \item{\code{geojson.coordinates}}{list COLUMN_DESCRIPTION}
  \item{\code{geojson.type}}{character COLUMN_DESCRIPTION}
  \item{\code{user.site_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.created_at}}{character COLUMN_DESCRIPTION}
  \item{\code{user.id}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.login}}{character COLUMN_DESCRIPTION}
  \item{\code{user.spam}}{logical COLUMN_DESCRIPTION}
  \item{\code{user.suspended}}{logical COLUMN_DESCRIPTION}
  \item{\code{user.login_autocomplete}}{character COLUMN_DESCRIPTION}
  \item{\code{user.login_exact}}{character COLUMN_DESCRIPTION}
  \item{\code{user.name}}{character COLUMN_DESCRIPTION}
  \item{\code{user.name_autocomplete}}{character COLUMN_DESCRIPTION}
  \item{\code{user.orcid}}{character COLUMN_DESCRIPTION}
  \item{\code{user.icon}}{character COLUMN_DESCRIPTION}
  \item{\code{user.observations_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.identifications_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.journal_posts_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.activity_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.species_count}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.universal_search_rank}}{integer COLUMN_DESCRIPTION}
  \item{\code{user.roles}}{list COLUMN_DESCRIPTION}
  \item{\code{user.icon_url}}{character COLUMN_DESCRIPTION}
  \item{\code{user.preferences.prefers_community_taxa}}{logical COLUMN_DESCRIPTION}
  \item{\code{user.preferences.prefers_observation_fields_by}}{character COLUMN_DESCRIPTION}
  \item{\code{user.preferences.prefers_project_addition_by}}{character COLUMN_DESCRIPTION}
  \item{\code{prov}}{character COLUMN_DESCRIPTION}
  \item{\code{longitude}}{character COLUMN_DESCRIPTION}
  \item{\code{latitude}}{character COLUMN_DESCRIPTION}
}
}
\usage{
inat_10k
}
\description{
A medium size dataset for testing
}
\details{
DETAILS
}
\keyword{datasets}
