% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_obis.R
\docType{data}
\name{obis_10k}
\alias{obis_10k}
\title{OBIS data| Heterogeneous | 10,000 record}
\format{
A data frame with 10000 rows and 225 variables:
\describe{
  \item{\code{rightsHolder}}{character COLUMN_DESCRIPTION}
  \item{\code{infraphylum}}{character COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{date_year}}{integer COLUMN_DESCRIPTION}
  \item{\code{scientificNameID}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{individualCount}}{character COLUMN_DESCRIPTION}
  \item{\code{dropped}}{logical COLUMN_DESCRIPTION}
  \item{\code{gigaclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{aphiaID}}{integer COLUMN_DESCRIPTION}
  \item{\code{latitude}}{double COLUMN_DESCRIPTION}
  \item{\code{subclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{type}}{character COLUMN_DESCRIPTION}
  \item{\code{gigaclass}}{character COLUMN_DESCRIPTION}
  \item{\code{infraphylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{phylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{familyid}}{integer COLUMN_DESCRIPTION}
  \item{\code{catalogNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{terrestrial}}{logical COLUMN_DESCRIPTION}
  \item{\code{id}}{character COLUMN_DESCRIPTION}
  \item{\code{parvphylum}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{recordNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{dataset_id}}{character COLUMN_DESCRIPTION}
  \item{\code{locality}}{character COLUMN_DESCRIPTION}
  \item{\code{longitude}}{double COLUMN_DESCRIPTION}
  \item{\code{collectionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{date_end}}{double COLUMN_DESCRIPTION}
  \item{\code{speciesid}}{integer COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{character COLUMN_DESCRIPTION}
  \item{\code{license}}{character COLUMN_DESCRIPTION}
  \item{\code{date_start}}{double COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventDate}}{double COLUMN_DESCRIPTION}
  \item{\code{brackish}}{logical COLUMN_DESCRIPTION}
  \item{\code{superorder}}{character COLUMN_DESCRIPTION}
  \item{\code{absence}}{logical COLUMN_DESCRIPTION}
  \item{\code{subfamily}}{character COLUMN_DESCRIPTION}
  \item{\code{superorderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{genusid}}{integer COLUMN_DESCRIPTION}
  \item{\code{originalScientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{marine}}{logical COLUMN_DESCRIPTION}
  \item{\code{subphylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{subfamilyid}}{integer COLUMN_DESCRIPTION}
  \item{\code{institutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{date_mid}}{double COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{orderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{recordedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{classid}}{integer COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{character COLUMN_DESCRIPTION}
  \item{\code{subphylum}}{character COLUMN_DESCRIPTION}
  \item{\code{subclass}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{kingdomid}}{integer COLUMN_DESCRIPTION}
  \item{\code{parvphylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{node_id}}{list COLUMN_DESCRIPTION}
  \item{\code{flags}}{list COLUMN_DESCRIPTION}
  \item{\code{sss}}{double COLUMN_DESCRIPTION}
  \item{\code{shoredistance}}{integer COLUMN_DESCRIPTION}
  \item{\code{sst}}{double COLUMN_DESCRIPTION}
  \item{\code{bathymetry}}{double COLUMN_DESCRIPTION}
  \item{\code{year}}{character COLUMN_DESCRIPTION}
  \item{\code{fieldNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{language}}{character COLUMN_DESCRIPTION}
  \item{\code{superclass}}{character COLUMN_DESCRIPTION}
  \item{\code{modified}}{character COLUMN_DESCRIPTION}
  \item{\code{maximumDepthInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{day}}{character COLUMN_DESCRIPTION}
  \item{\code{superclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{footprintSRS}}{character COLUMN_DESCRIPTION}
  \item{\code{month}}{character COLUMN_DESCRIPTION}
  \item{\code{samplingProtocol}}{character COLUMN_DESCRIPTION}
  \item{\code{parentEventID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{minimumDepthInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{preparations}}{character COLUMN_DESCRIPTION}
  \item{\code{infraclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{infraclass}}{character COLUMN_DESCRIPTION}
  \item{\code{sex}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetName}}{character COLUMN_DESCRIPTION}
  \item{\code{geodeticDatum}}{character COLUMN_DESCRIPTION}
  \item{\code{depth}}{double COLUMN_DESCRIPTION}
  \item{\code{lifeStage}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetID}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionID}}{character COLUMN_DESCRIPTION}
  \item{\code{superfamilyid}}{integer COLUMN_DESCRIPTION}
  \item{\code{infraorderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{superfamily}}{character COLUMN_DESCRIPTION}
  \item{\code{suborderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{scientificNameAuthorship}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonRank}}{character COLUMN_DESCRIPTION}
  \item{\code{vernacularName}}{character COLUMN_DESCRIPTION}
  \item{\code{locationRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{suborder}}{character COLUMN_DESCRIPTION}
  \item{\code{infraorder}}{character COLUMN_DESCRIPTION}
  \item{\code{dataGeneralizations}}{character COLUMN_DESCRIPTION}
  \item{\code{specificEpithet}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedReferences}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferencedDate}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimEventDate}}{character COLUMN_DESCRIPTION}
  \item{\code{organismID}}{character COLUMN_DESCRIPTION}
  \item{\code{dateIdentified}}{character COLUMN_DESCRIPTION}
  \item{\code{ownerInstitutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{bibliographicCitation}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateUncertaintyInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{eventTime}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{nomenclaturalCode}}{character COLUMN_DESCRIPTION}
  \item{\code{footprintWKT}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonomicStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{waterBody}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinatePrecision}}{character COLUMN_DESCRIPTION}
  \item{\code{organismRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{locationID}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimCoordinates}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{character COLUMN_DESCRIPTION}
  \item{\code{habitat}}{character COLUMN_DESCRIPTION}
  \item{\code{references}}{character COLUMN_DESCRIPTION}
  \item{\code{minimumElevationInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{stateProvince}}{character COLUMN_DESCRIPTION}
  \item{\code{maximumElevationInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{samplingEffort}}{character COLUMN_DESCRIPTION}
  \item{\code{dynamicProperties}}{character COLUMN_DESCRIPTION}
  \item{\code{startDayOfYear}}{character COLUMN_DESCRIPTION}
  \item{\code{subkingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{infrakingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{subkingdomid}}{integer COLUMN_DESCRIPTION}
  \item{\code{infrakingdomid}}{integer COLUMN_DESCRIPTION}
  \item{\code{tribeid}}{integer COLUMN_DESCRIPTION}
  \item{\code{tribe}}{character COLUMN_DESCRIPTION}
  \item{\code{municipality}}{character COLUMN_DESCRIPTION}
  \item{\code{countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{sampleSizeUnit}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferencedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{island}}{character COLUMN_DESCRIPTION}
  \item{\code{islandGroup}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonID}}{character COLUMN_DESCRIPTION}
  \item{\code{identifiedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferenceProtocol}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimDepth}}{character COLUMN_DESCRIPTION}
  \item{\code{sampleSizeValue}}{character COLUMN_DESCRIPTION}
  \item{\code{organismQuantity}}{character COLUMN_DESCRIPTION}
  \item{\code{organismQuantityType}}{character COLUMN_DESCRIPTION}
  \item{\code{informationWithheld}}{character COLUMN_DESCRIPTION}
  \item{\code{materialSampleID}}{character COLUMN_DESCRIPTION}
  \item{\code{subgenus}}{character COLUMN_DESCRIPTION}
  \item{\code{subgenusid}}{integer COLUMN_DESCRIPTION}
  \item{\code{continent}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationQualifier}}{character COLUMN_DESCRIPTION}
  \item{\code{hab}}{logical COLUMN_DESCRIPTION}
  \item{\code{maximumDistanceAboveSurfaceInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferenceRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{locationAccordingTo}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimLocality}}{character COLUMN_DESCRIPTION}
  \item{\code{rights}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedMedia}}{character COLUMN_DESCRIPTION}
  \item{\code{wrims}}{logical COLUMN_DESCRIPTION}
  \item{\code{originalNameUsage}}{character COLUMN_DESCRIPTION}
  \item{\code{parvorder}}{character COLUMN_DESCRIPTION}
  \item{\code{parvorderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{acceptedNameUsage}}{character COLUMN_DESCRIPTION}
  \item{\code{acceptedNameUsageID}}{character COLUMN_DESCRIPTION}
  \item{\code{accessRights}}{character COLUMN_DESCRIPTION}
  \item{\code{county}}{character COLUMN_DESCRIPTION}
  \item{\code{nameAccordingTo}}{character COLUMN_DESCRIPTION}
  \item{\code{behavior}}{character COLUMN_DESCRIPTION}
  \item{\code{otherCatalogNumbers}}{character COLUMN_DESCRIPTION}
  \item{\code{subspecies}}{character COLUMN_DESCRIPTION}
  \item{\code{subspeciesid}}{integer COLUMN_DESCRIPTION}
  \item{\code{organismScope}}{character COLUMN_DESCRIPTION}
  \item{\code{section}}{character COLUMN_DESCRIPTION}
  \item{\code{subsection}}{character COLUMN_DESCRIPTION}
  \item{\code{sectionid}}{integer COLUMN_DESCRIPTION}
  \item{\code{subsectionid}}{integer COLUMN_DESCRIPTION}
  \item{\code{taxonConceptID}}{character COLUMN_DESCRIPTION}
  \item{\code{higherGeography}}{character COLUMN_DESCRIPTION}
  \item{\code{endDayOfYear}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimLatitude}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimLongitude}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimCoordinateSystem}}{character COLUMN_DESCRIPTION}
  \item{\code{disposition}}{character COLUMN_DESCRIPTION}
  \item{\code{reproductiveCondition}}{character COLUMN_DESCRIPTION}
  \item{\code{establishmentMeans}}{character COLUMN_DESCRIPTION}
  \item{\code{higherClassification}}{character COLUMN_DESCRIPTION}
  \item{\code{typeStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferenceVerificationStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{fieldNotes}}{character COLUMN_DESCRIPTION}
  \item{\code{nameAccordingToID}}{character COLUMN_DESCRIPTION}
  \item{\code{previousIdentifications}}{character COLUMN_DESCRIPTION}
  \item{\code{minimumDistanceAboveSurfaceInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{namePublishedIn}}{character COLUMN_DESCRIPTION}
  \item{\code{subterclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{subterclass}}{character COLUMN_DESCRIPTION}
  \item{\code{infraspecificEpithet}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimSRS}}{character COLUMN_DESCRIPTION}
  \item{\code{subtribeid}}{integer COLUMN_DESCRIPTION}
  \item{\code{subtribe}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationVerificationStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{NA}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationID}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedSequences}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferenceSources}}{character COLUMN_DESCRIPTION}
  \item{\code{organismName}}{character COLUMN_DESCRIPTION}
  \item{\code{namePublishedInID}}{character COLUMN_DESCRIPTION}
  \item{\code{higherGeographyID}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationReferences}}{character COLUMN_DESCRIPTION}
  \item{\code{forma}}{character COLUMN_DESCRIPTION}
  \item{\code{formaid}}{integer COLUMN_DESCRIPTION}
  \item{\code{identifiedByID}}{character COLUMN_DESCRIPTION}
  \item{\code{recordedByID}}{character COLUMN_DESCRIPTION}
  \item{\code{varietyid}}{integer COLUMN_DESCRIPTION}
  \item{\code{variety}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimTaxonRank}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedTaxa}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedOccurrences}}{character COLUMN_DESCRIPTION}
  \item{\code{prov}}{character COLUMN_DESCRIPTION}
}
}
\usage{
obis_10k
}
\description{
A medium size dataset for testing
}
\details{
DETAILS
}
\keyword{datasets}
