% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_obis.R
\docType{data}
\name{obis_30r}
\alias{obis_30r}
\title{OBIS data | Heterogeneous | 30 record}
\format{
A data frame with 30 rows and 154 variables:
\describe{
  \item{\code{rightsHolder}}{character COLUMN_DESCRIPTION}
  \item{\code{infraphylum}}{character COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{date_year}}{integer COLUMN_DESCRIPTION}
  \item{\code{scientificNameID}}{character COLUMN_DESCRIPTION}
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{individualCount}}{character COLUMN_DESCRIPTION}
  \item{\code{dropped}}{logical COLUMN_DESCRIPTION}
  \item{\code{gigaclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{aphiaID}}{integer COLUMN_DESCRIPTION}
  \item{\code{latitude}}{double COLUMN_DESCRIPTION}
  \item{\code{subclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{type}}{character COLUMN_DESCRIPTION}
  \item{\code{gigaclass}}{character COLUMN_DESCRIPTION}
  \item{\code{infraphylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{phylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{familyid}}{integer COLUMN_DESCRIPTION}
  \item{\code{catalogNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{terrestrial}}{logical COLUMN_DESCRIPTION}
  \item{\code{id}}{character COLUMN_DESCRIPTION}
  \item{\code{parvphylum}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{recordNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{dataset_id}}{character COLUMN_DESCRIPTION}
  \item{\code{locality}}{character COLUMN_DESCRIPTION}
  \item{\code{longitude}}{double COLUMN_DESCRIPTION}
  \item{\code{collectionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{date_end}}{double COLUMN_DESCRIPTION}
  \item{\code{speciesid}}{integer COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{character COLUMN_DESCRIPTION}
  \item{\code{license}}{character COLUMN_DESCRIPTION}
  \item{\code{date_start}}{double COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventDate}}{double COLUMN_DESCRIPTION}
  \item{\code{brackish}}{logical COLUMN_DESCRIPTION}
  \item{\code{superorder}}{character COLUMN_DESCRIPTION}
  \item{\code{absence}}{logical COLUMN_DESCRIPTION}
  \item{\code{subfamily}}{character COLUMN_DESCRIPTION}
  \item{\code{superorderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{genusid}}{integer COLUMN_DESCRIPTION}
  \item{\code{originalScientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{marine}}{logical COLUMN_DESCRIPTION}
  \item{\code{subphylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{subfamilyid}}{integer COLUMN_DESCRIPTION}
  \item{\code{institutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{date_mid}}{double COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{orderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{recordedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{classid}}{integer COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{character COLUMN_DESCRIPTION}
  \item{\code{subphylum}}{character COLUMN_DESCRIPTION}
  \item{\code{subclass}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{kingdomid}}{integer COLUMN_DESCRIPTION}
  \item{\code{parvphylumid}}{integer COLUMN_DESCRIPTION}
  \item{\code{node_id}}{list COLUMN_DESCRIPTION}
  \item{\code{flags}}{list COLUMN_DESCRIPTION}
  \item{\code{sss}}{double COLUMN_DESCRIPTION}
  \item{\code{shoredistance}}{integer COLUMN_DESCRIPTION}
  \item{\code{sst}}{double COLUMN_DESCRIPTION}
  \item{\code{bathymetry}}{double COLUMN_DESCRIPTION}
  \item{\code{year}}{character COLUMN_DESCRIPTION}
  \item{\code{fieldNumber}}{character COLUMN_DESCRIPTION}
  \item{\code{language}}{character COLUMN_DESCRIPTION}
  \item{\code{superclass}}{character COLUMN_DESCRIPTION}
  \item{\code{modified}}{character COLUMN_DESCRIPTION}
  \item{\code{maximumDepthInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{day}}{character COLUMN_DESCRIPTION}
  \item{\code{superclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{footprintSRS}}{character COLUMN_DESCRIPTION}
  \item{\code{month}}{character COLUMN_DESCRIPTION}
  \item{\code{samplingProtocol}}{character COLUMN_DESCRIPTION}
  \item{\code{parentEventID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventID}}{character COLUMN_DESCRIPTION}
  \item{\code{eventRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{minimumDepthInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{preparations}}{character COLUMN_DESCRIPTION}
  \item{\code{infraclassid}}{integer COLUMN_DESCRIPTION}
  \item{\code{infraclass}}{character COLUMN_DESCRIPTION}
  \item{\code{sex}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetName}}{character COLUMN_DESCRIPTION}
  \item{\code{geodeticDatum}}{character COLUMN_DESCRIPTION}
  \item{\code{depth}}{double COLUMN_DESCRIPTION}
  \item{\code{lifeStage}}{character COLUMN_DESCRIPTION}
  \item{\code{datasetID}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionID}}{character COLUMN_DESCRIPTION}
  \item{\code{superfamilyid}}{integer COLUMN_DESCRIPTION}
  \item{\code{infraorderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{superfamily}}{character COLUMN_DESCRIPTION}
  \item{\code{suborderid}}{integer COLUMN_DESCRIPTION}
  \item{\code{scientificNameAuthorship}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonRank}}{character COLUMN_DESCRIPTION}
  \item{\code{vernacularName}}{character COLUMN_DESCRIPTION}
  \item{\code{locationRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{suborder}}{character COLUMN_DESCRIPTION}
  \item{\code{infraorder}}{character COLUMN_DESCRIPTION}
  \item{\code{dataGeneralizations}}{character COLUMN_DESCRIPTION}
  \item{\code{specificEpithet}}{character COLUMN_DESCRIPTION}
  \item{\code{associatedReferences}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferencedDate}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimEventDate}}{character COLUMN_DESCRIPTION}
  \item{\code{organismID}}{character COLUMN_DESCRIPTION}
  \item{\code{dateIdentified}}{character COLUMN_DESCRIPTION}
  \item{\code{ownerInstitutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{bibliographicCitation}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateUncertaintyInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{eventTime}}{character COLUMN_DESCRIPTION}
  \item{\code{identificationRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{nomenclaturalCode}}{character COLUMN_DESCRIPTION}
  \item{\code{footprintWKT}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonomicStatus}}{character COLUMN_DESCRIPTION}
  \item{\code{waterBody}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinatePrecision}}{character COLUMN_DESCRIPTION}
  \item{\code{organismRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{locationID}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimCoordinates}}{character COLUMN_DESCRIPTION}
  \item{\code{category}}{character COLUMN_DESCRIPTION}
  \item{\code{habitat}}{character COLUMN_DESCRIPTION}
  \item{\code{references}}{character COLUMN_DESCRIPTION}
  \item{\code{minimumElevationInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{stateProvince}}{character COLUMN_DESCRIPTION}
  \item{\code{maximumElevationInMeters}}{character COLUMN_DESCRIPTION}
  \item{\code{samplingEffort}}{character COLUMN_DESCRIPTION}
  \item{\code{dynamicProperties}}{character COLUMN_DESCRIPTION}
  \item{\code{startDayOfYear}}{character COLUMN_DESCRIPTION}
  \item{\code{subkingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{infrakingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{subkingdomid}}{integer COLUMN_DESCRIPTION}
  \item{\code{infrakingdomid}}{integer COLUMN_DESCRIPTION}
  \item{\code{tribeid}}{integer COLUMN_DESCRIPTION}
  \item{\code{tribe}}{character COLUMN_DESCRIPTION}
  \item{\code{municipality}}{character COLUMN_DESCRIPTION}
  \item{\code{countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{sampleSizeUnit}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferencedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{island}}{character COLUMN_DESCRIPTION}
  \item{\code{islandGroup}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonID}}{character COLUMN_DESCRIPTION}
  \item{\code{identifiedBy}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferenceProtocol}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimDepth}}{character COLUMN_DESCRIPTION}
  \item{\code{sampleSizeValue}}{character COLUMN_DESCRIPTION}
  \item{\code{organismQuantity}}{character COLUMN_DESCRIPTION}
  \item{\code{organismQuantityType}}{character COLUMN_DESCRIPTION}
  \item{\code{informationWithheld}}{character COLUMN_DESCRIPTION}
  \item{\code{prov}}{character COLUMN_DESCRIPTION}
}
}
\usage{
obis_30r
}
\description{
A small and heterogeneous dataset for testing
}
\details{
DETAILS
}
\keyword{datasets}
