% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vertnet.R
\docType{data}
\name{vertnet_30r}
\alias{vertnet_30r}
\title{VertNet data | Heterogeneous |30 record}
\format{
A data frame with 30 rows and 51 variables:
\describe{
  \item{\code{name}}{character COLUMN_DESCRIPTION}
  \item{\code{longitude}}{character COLUMN_DESCRIPTION}
  \item{\code{latitude}}{character COLUMN_DESCRIPTION}
  \item{\code{prov}}{character COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{recordedby}}{character COLUMN_DESCRIPTION}
  \item{\code{higherclassification}}{character COLUMN_DESCRIPTION}
  \item{\code{basisofrecord}}{character COLUMN_DESCRIPTION}
  \item{\code{month}}{character COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{references}}{character COLUMN_DESCRIPTION}
  \item{\code{year}}{character COLUMN_DESCRIPTION}
  \item{\code{startdayofyear}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonrank}}{character COLUMN_DESCRIPTION}
  \item{\code{specificepithet}}{character COLUMN_DESCRIPTION}
  \item{\code{bibliographiccitation}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{countrycode}}{character COLUMN_DESCRIPTION}
  \item{\code{geodeticdatum}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateuncertaintyinmeters}}{character COLUMN_DESCRIPTION}
  \item{\code{highergeography}}{character COLUMN_DESCRIPTION}
  \item{\code{accessrights}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimlocality}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimeventdate}}{character COLUMN_DESCRIPTION}
  \item{\code{day}}{character COLUMN_DESCRIPTION}
  \item{\code{eventid}}{character COLUMN_DESCRIPTION}
  \item{\code{collectioncode}}{character COLUMN_DESCRIPTION}
  \item{\code{islandgroup}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrencestatus}}{character COLUMN_DESCRIPTION}
  \item{\code{locationremarks}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateprecision}}{character COLUMN_DESCRIPTION}
  \item{\code{institutioncode}}{character COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceid}}{character COLUMN_DESCRIPTION}
  \item{\code{language}}{character COLUMN_DESCRIPTION}
  \item{\code{license}}{character COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{georeferenceverificationstatus}}{character COLUMN_DESCRIPTION}
  \item{\code{modified}}{character COLUMN_DESCRIPTION}
  \item{\code{eventdate}}{double COLUMN_DESCRIPTION}
  \item{\code{nomenclaturalcode}}{character COLUMN_DESCRIPTION}
  \item{\code{continent}}{character COLUMN_DESCRIPTION}
  \item{\code{waterbody}}{character COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{catalognumber}}{character COLUMN_DESCRIPTION}
  \item{\code{enddayofyear}}{character COLUMN_DESCRIPTION}
  \item{\code{stateprovince}}{character COLUMN_DESCRIPTION}
  \item{\code{vernacularname}}{character COLUMN_DESCRIPTION}
  \item{\code{identifiedby}}{character COLUMN_DESCRIPTION}
  \item{\code{dateidentified}}{character COLUMN_DESCRIPTION}
}
}
\usage{
vertnet_30r
}
\description{
A small and heterogeneous dataset for testing
}
\details{
DETAILS
}
\keyword{datasets}
